+++ src/main/java/com/dotweblabs/gwt/client/ParseResponse.java
@@ -1,5 +1,22 @@
+/**
+ *
+ * Copyright (c) 2016 Dotweblabs Web Technologies and others. All rights reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * 	   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
 package com.dotweblabs.gwt.client;
 
+import com.google.gwt.json.client.JSONArray;
 import com.google.gwt.json.client.JSONObject;
 import com.google.gwt.json.client.JSONParser;
 import com.google.gwt.json.client.JSONValue;
@@ -8,7 +25,11 @@ import java.util.HashMap;
 import java.util.Iterator;
 
 /**
- * Created by paperspace on 6/29/2016.
+ *
+ *
+ * @author Kerby Martino
+ * @since 0-SNAPSHOT
+ * @version 0-SNAPSHOT
  */
 public class ParseResponse extends JSONObject {
     public static ParseResponse parse(String json) {
@@ -45,5 +66,9 @@ public class ParseResponse extends JSONObject {
             return get("updatedAt").isString().stringValue();
         }
         return null;
+    }
+    public JSONArray getResults() {
+        JSONArray results =  get("results") != null ? get("results").isArray() : null;
+       return results;
     }
 }